+++ build.ps1
@@ -3,6 +3,5 @@ pyinstaller main.py `
   --onefile `
   --noconsole `
   --add-data "assets;assets" `
-  --add-data "installers/dotnet-install.ps1;installers" `
-  --add-data "installers/MelonLoader.Installer.exe;installers" `
+  --add-data "assets/MelonLoader.x64.zip;assets" `
   --icon "assets\DIP_Installer_Icon.png"