+++ build.sh
@@ -5,6 +5,5 @@ pyinstaller main.py \
   --onefile \
   --noconsole \
   --add-data "assets:assets" \
-  --add-data "installers/dotnet-install.sh:installers" \
-  --add-data "installers/MelonLoader.Installer.Linux:installers" \
+  --add-data "assets/MelonLoader.x64.zip:assets" \
   --icon assets/DIP_Installer_Icon.png